+++ test/Tree.ts
@@ -5,7 +5,7 @@ import * as O from '../src/Option'
 import * as _ from '../src/Tree'
 import * as U from './util'
 import * as N from '../src/number'
-import * as I from '../src/Identity'
+import * as T from '../src/Task'
 
 describe('Tree', () => {
   // -------------------------------------------------------------------------------------
@@ -111,6 +111,21 @@ describe('Tree', () => {
     )
   })
 
+  it('unfoldTreeM', async () => {
+    U.deepStrictEqual(
+      _.unfoldTreeM({ ...O.Monad, ...O.Applicative })(1, (b) => O.some([b, b < 3 ? [b + 1, b + 2] : []])),
+      O.some(_.make(1, [_.make(2, [_.make(3), _.make(4)]), _.make(3)]))
+    )
+    U.deepStrictEqual(
+      _.unfoldTreeM({ ...O.Monad, ...O.Applicative })(1, (b) => (b < 3 ? O.some([b, [b + 1, b + 2]]) : O.none)),
+      O.none
+    )
+    U.deepStrictEqual(
+      await _.unfoldTreeM({ ...T.Monad, ...T.ApplicativePar })(1, (b) => T.of([b, b < 3 ? [b + 1, b + 2] : []]))(),
+      _.make(1, [_.make(2, [_.make(3), _.make(4)]), _.make(3)])
+    )
+  })
+
   // -------------------------------------------------------------------------------------
   // instances
   // -------------------------------------------------------------------------------------
@@ -143,12 +158,6 @@ describe('Tree', () => {
       _.fold((a: number, bs: ReadonlyArray<number>) => bs.reduce((b, acc) => Math.max(b, acc), a))(t),
       3
     )
-  })
-
-  it('unfoldTreeM', () => {
-    const fa = _.unfoldTreeM({ ...I.Monad, ...I.Applicative })(1, (b) => [b, b < 3 ? [b + 1, b + 2] : []])
-    const expected = _.make(1, [_.make(2, [_.make(3), _.make(4)]), _.make(3)])
-    U.deepStrictEqual(fa, expected)
   })
 
   // -------------------------------------------------------------------------------------