+++ CHANGELOG.md
@@ -15,6 +15,11 @@
 **Note**: A feature tagged as Experimental is in a
 high state of flux, you're at risk of it changing without notice.
 
+# 2.9.4
+
+- **Bug Fix**
+  - fix `filter` overloads in `NonEmptyArray` / `ReadonlyNonEmptyArray`, closes #1388 (@gcanti)
+
 # 2.9.3
 
 - **Polish**